Rename filterschedule package to tracksfilter
