+++ app/src/main/java/net/squanchy/injection/ApplicationComponent.kt
@@ -6,8 +6,8 @@ import net.squanchy.analytics.Analytics
 import net.squanchy.analytics.AnalyticsModule
 import net.squanchy.remoteconfig.RemoteConfig
 import net.squanchy.remoteconfig.RemoteConfigModule
-import net.squanchy.schedule.filterschedule.TracksFilter
-import net.squanchy.schedule.filterschedule.TracksFilterModule
+import net.squanchy.schedule.tracksfilter.TracksFilter
+import net.squanchy.schedule.tracksfilter.TracksFilterModule
 import net.squanchy.service.firebase.FirebaseAuthService
 import net.squanchy.service.firebase.FirebaseDbService
 import net.squanchy.service.firebase.injection.FirebaseModule
