+++ app/src/main/java/net/squanchy/schedule/ScheduleService.kt
@@ -5,7 +5,7 @@ import io.reactivex.functions.BiFunction
 import net.squanchy.schedule.domain.view.Schedule
 import net.squanchy.schedule.domain.view.SchedulePage
 import net.squanchy.schedule.domain.view.Track
-import net.squanchy.schedule.filterschedule.TracksFilter
+import net.squanchy.schedule.tracksfilter.TracksFilter
 import net.squanchy.service.firebase.FirebaseAuthService
 import net.squanchy.service.firestore.FirestoreDbService
 import net.squanchy.service.firestore.model.schedule.FirestoreEvent
