+++ app/src/main/java/net/squanchy/schedule/tracksfilter/TracksFilterComponent.kt
@@ -1,4 +1,4 @@
-package net.squanchy.schedule.filterschedule
+package net.squanchy.schedule.tracksfilter
 
 import android.content.Context
 import dagger.Component
