+++ app/src/main/java/net/squanchy/schedule/tracksfilter/TracksFilterModule.kt
@@ -1,4 +1,4 @@
-package net.squanchy.schedule.filterschedule
+package net.squanchy.schedule.tracksfilter
 
 import dagger.Module
 import dagger.Provides
